Proof of concept of delegating monitoring to prmon
Show how prmon can be used as an external resource monitor for
jMonitor
- prmon is invoked after the main process is spawned and monitors it
  and all of its children
- the prmon "snapshot" file is read on each monitoring cycle to read
  current resource values
- these values are pushed into the current jMonitor VirtMonitor classes
  (using a new method, push())

This only a proof of concept to show the idea, I wouldn't use this code
in its current form.

The content of the current prmon snapshot file isn't quite a match to
what jMonitor uses, but it could rather easily be ammended (the final
prmon txt file does contain everything needed, but it's much more
awkward to parse on the fly).

In particular:
- prmon doesn't measure CPU rate, but instead measures cummulative
  consuption time (user + system) so for now this has been used as a
  proxy in the CPURate monitor.
- the VMEM and RSS values are strictly the Max values at the moment,
  so they are monotonic increasing; prmon could easily output the
  correct snapshot values if needed